--- a/runtime/caml/domain.h
+++ b/runtime/caml/domain.h
@@ -100,6 +100,7 @@ int caml_domain_is_in_stw(void);
 #endif
 
 int caml_try_run_on_all_domains_with_spin_work(
+  int sync,
   void (*handler)(caml_domain_state*, void*, int, caml_domain_state**),
   void* data,
   void (*leader_setup)(caml_domain_state*),