--- a/runtime/caml/domain_state.tbl
+++ b/runtime/caml/domain_state.tbl
@@ -92,6 +92,7 @@ DOMAIN_STATE(intnat, compare_unordered)
 DOMAIN_STATE(uintnat, oo_next_id_local)
 
 DOMAIN_STATE(uintnat, requested_major_slice)
+DOMAIN_STATE(uintnat, requested_global_major_slice)
 
 DOMAIN_STATE(uintnat, requested_minor_gc)