--- a/runtime/caml/domain_state.tbl
+++ b/runtime/caml/domain_state.tbl
@@ -118,10 +118,7 @@ DOMAIN_STATE(double, extra_heap_resources_minor)
 DOMAIN_STATE(uintnat, dependent_size)
 DOMAIN_STATE(uintnat, dependent_allocated)
 
-DOMAIN_STATE(double, alloc_ratio)
-DOMAIN_STATE(double, dependent_ratio)
-DOMAIN_STATE(double, extra_ratio)
-DOMAIN_STATE(intnat, slice_budget)
+DOMAIN_STATE(intnat, slice_target)
 DOMAIN_STATE(intnat, major_work_done_between_slices)
 
 DOMAIN_STATE(struct caml_extern_state*, extern_state)