--- a/runtime/caml/domain_state.tbl
+++ b/runtime/caml/domain_state.tbl
@@ -122,7 +122,6 @@ DOMAIN_STATE(double, alloc_ratio)
 DOMAIN_STATE(double, dependent_ratio)
 DOMAIN_STATE(double, extra_ratio)
 DOMAIN_STATE(intnat, slice_budget)
-DOMAIN_STATE(int, is_actively_collecting)
 
 DOMAIN_STATE(struct caml_extern_state*, extern_state)
 DOMAIN_STATE(struct caml_intern_state*, intern_state)