--- a/runtime/domain.c
+++ b/runtime/domain.c
@@ -602,6 +602,8 @@ static void domain_create(uintnat initial_minor_heap_wsize) {
   domain_state->dependent_size = 0;
   domain_state->dependent_allocated = 0;
 
+  domain_state->major_work_done_between_slices = 0;
+
   /* the minor heap will be initialized by
      [caml_reallocate_minor_heap] below. */
   domain_state->young_start = NULL;
@@ -948,11 +950,9 @@ static void* backup_thread_func(void* v)
          *  - need to guarantee no blocking so that backup thread
          *    can be signalled from caml_leave_blocking_section
          */
-        if (caml_incoming_interrupts_queued()
-            //|| caml_check_gc_interrupt(caml_state)
-            ) {
+        if (caml_incoming_interrupts_queued()) {
           if (caml_plat_try_lock(&di->domain_lock)) {
-            caml_handle_incoming_interrupts(); // caml_handle_gc_interrupt();
+            caml_handle_incoming_interrupts();
             caml_plat_unlock(&di->domain_lock);
           }
         }
@@ -962,9 +962,7 @@ static void* backup_thread_func(void* v)
         caml_plat_lock(&s->lock);
         msg = atomic_load_acquire (&di->backup_thread_msg);
         if (msg == BT_IN_BLOCKING_SECTION &&
-            !caml_incoming_interrupts_queued()
-            // && !caml_check_gc_interrupt(caml_state)
-            )
+            !caml_incoming_interrupts_queued())
           caml_plat_wait(&s->cond);
         caml_plat_unlock(&s->lock);
         break;
@@ -1578,12 +1576,14 @@ Caml_inline void advance_global_major_slice_epoch (caml_domain_state* d)
   }
 }
 
-static void global_major_slice_stw_handler (caml_domain_state *domain,
-                                            void *unused,
-                                            int participating_count,
-                                            caml_domain_state **participating)
+static void global_major_slice_callback (caml_domain_state *domain,
+                                         void *unused,
+                                         int participating_count,
+                                         caml_domain_state **participating)
 {
   domain->requested_major_slice = 1;
+  /* Nothing else to do, as [stw_hander] will call [caml_poll_gc_work]
+     right after the callback. */
 }
 
 void caml_poll_gc_work(void)
@@ -1635,7 +1635,7 @@ void caml_poll_gc_work(void)
   if (d->requested_global_major_slice) {
     d->requested_global_major_slice = 0;
     (void) caml_try_run_on_all_domains_async(
-             &global_major_slice_stw_handler, NULL, NULL);
+             &global_major_slice_callback, NULL, NULL);
   }
 
   if (atomic_load_acquire(&d->requested_external_interrupt)) {