--- a/runtime/memory.c
+++ b/runtime/memory.c
@@ -195,7 +195,7 @@ CAMLexport void caml_adjust_gc_speed (mlsize_t res, mlsize_t max)
   if (Caml_state->extra_heap_resources > 1.0){
     CAML_EV_COUNTER (EV_C_REQUEST_MAJOR_ADJUST_GC_SPEED, 1);
     Caml_state->extra_heap_resources = 1.0;
-    caml_request_major_slice ();
+    caml_request_major_slice_global ();
   }
 }
 
@@ -344,7 +344,7 @@ Caml_inline value alloc_shr(mlsize_t wosize, tag_t tag, reserved_t reserved,
   dom_st->allocated_words += Whsize_wosize(wosize);
   if (dom_st->allocated_words > dom_st->minor_heap_wsz) {
     CAML_EV_COUNTER (EV_C_REQUEST_MAJOR_ALLOC_SHR, 1);
-    caml_request_major_slice();
+    caml_request_major_slice_global();
   }
 
 #ifdef DEBUG