--- a/testsuite/tests/lib-runtime-events/test_instrumented.reference
+++ b/testsuite/tests/lib-runtime-events/test_instrumented.reference
@@ -1,1 +1,1 @@
-lost_event_words: 0, total_sizes: 2000004, total_minors: 30
+lost_event_words: 0, total_sizes: 2000004, total_minors: 31