Trying to fix #11589 (and maybe #11548). Two changes here:

1. GC work accounting and distribution becomes global.
2. extra major slices are done globally instead of locally.

The benchmark code is here: https://gist.github.com/damiendoligez/86a3afa6c6a594e0274d6194d91457b7

### GC work accounting
Major allocations are counted globally and the counts are evenly distributed among active domains. These counts tell the major GC slices how much work they have to do. The main difficulty is that some domains may run out of major GC work for the current cycle before others. Then they cannot do any more GC work before the global synchronization at the end of the GC cycle (i.e. the STW event that starts the next cycle). In this case we have to redistribute their counts back to the domains that still have something to do.
This accounting tries to guarante that the global speed of major collection keeps up with the speed of allocation.

### extra major slices
When we allocate directly in the major heap, we sometimes have to trigger major GC slices in addition to the ones that are automatically run between minor collections. In trunk, this is done only in the allocating domain, but that's not enough: in the benchmark program, the allocating domain has already run out of GC work and these slices do nothing and allocation gets ahead of collection. In order to do these global slices, we have to send a signal to all other domains, but we don't need synchronization, do I modified the STW code to add an asynchronous mode.

### benchmark results
| test | max heap size (MB) | minor collections | major collections |
|-----|---------------|-------------------|------------------|
|trunk sequential | 30.5 | 125 | 116 |
|trunk with idle domain | 694 | 121 | 6 |
|this PR sequential | 52.6 | 112 | 112 |
|this PR with idle domain | 100 | 194 | 47 |

We can clearly see some improvement, but I still have to investigate:
- why the number of minor collections changes so much
- why the sequential test of this PR uses much more memory than trunk
- what is still holding up the major GC