@gasche 2023-02-28
What is the status of this PR? It looks like it does fix or at least greatly alleviate memory-consumption issues. Do we want to eventually get it merged? What do we need to do?