@damiendoligez 2023-02-28
I'm still working on this. I'm not satisfied with the increase of memory consumption in the sequential case.