@kayceesrk 2023-03-01
@damiendoligez I wondered if you had an idealised model for major slice scheduling in the multi-domain case. One thing we can do is to use [olly](https://github.com/tarides/runtime_events_tools#olly) to generate the traces and analyse run post-facto to see how close this PR gets to this ideal. [Perfetto](https://perfetto.dev/#analysis) has a SQL-based trace analysis framework where we can explore the trace by asking queries.  

IINM, runtime events currently does not report counter events for major heap sizes and estimated live data, which would be a pre-requisite for feeding into this idealised model. CC @sadiqj.