@damiendoligez 2023-03-02
I don't have a detailed model, I only know that, going back to https://gist.github.com/stedolan/52c13d6b1a30276db31ca98683c8db16, with the default overhead parameter we need to do about 5.2 units of work (the average of m and s) for each word allocated. When the GC does much less than that, the heap will get unacceptably large. The tricky part is distributing this work among the domains that do have some GC work available.