@damiendoligez 2023-04-11
I have much simplified the accounting logic: there is now a global counter of how much needs to be done, and each major slice tries to do that much work and decreases the global counter by how much work was actually done.

We still need to trigger major slices on other domains when the current domain runs out of things to do. This means we sometimes have all domains doing a major slice at the same time. There is a work-sharing mechanism to handle this case without doing N times the total amount of work.

The results are now better than trunk, even without any idle thread, but they still get worse when there are idle threads:

This PR:

| spawn | heap max (MB) | minor GC | major GC | work-to-alloc | backlog |
|-------|---------------|----------|----------|---------------|---------|
| 0     | 15.62         | 434      | 434      | 4.00          | 0.00    |
| 1     | 26.53         | 231      | 231      | 3.08          | 0.23    |
| 2     | 30.99         | 180      | 180      | 2.88          | 0.28    |
| 3     | 31.12         | 165      | 165      | 2.82          | 0.29    |
| 5     | 31.15         | 159      | 159      | 2.80          | 0.30    |
| 10    | 31.71         | 160      | 159      | 2.86          | 0.28    |


Trunk:

| spawn | heap max (MB) | minor GC | major GC | work-to-alloc |
|-------|---------------|----------|----------|---------------|
| 0     | 30.47         | 130      | 126      | 3.31          |

The work-to-alloc column gives the ratio of GC work done to number of words allocated in the major heap. When the GC's control loop works perfectly, this is 4 (assuming default GC settings). The backlog colum gives the amount of work that is left in the counter at the end of the program over the total amount of work.


I tried to increase the running time of the test with 0 or 1 idle thread:

no idle thread:

| steps | heap max (MB) | minor GC | major GC | work-to-alloc | backlog |
|-------|---------------|----------|----------|---------------|---------|
| 500   | 15.62         | 434      | 434      | 4.00          | 0.00    |
| 1000  | 30.27         | 918      | 918      | 4.00          | 0.00    |
| 1500  | 46.14         | 1409     | 1409     | 4.00          | 0.00    |
| 2000  | 61.40         | 1902     | 1902     | 4.00          | 0.00    |


one idle thread:

| steps | heap max (MB) | minor GC | major GC | work-to-alloc | backlog |
|-------|---------------|----------|----------|---------------|---------|
| 500   | 26.53         | 231      | 231      | 3.08          | 0.23    |
| 1000  | 54.92         | 473      | 473      | 3.03          | 0.24    |
| 1500  | 89.34         | 716      | 716      | 3.00          | 0.25    |
| 2000  | 103.00        | 960      | 960      | 3.01          | 0.25    |


The work-to-alloc ratio and residual backlog seem to be stable when the benchmark runs longer.


I still need to do a smoke check with sandmark and check whether this will fix #11548.