@gasche 2023-04-11
If I read your numbers right, in the sequential case (spawn=0) the benchmark runs many more GC slices in your PR than in trunk: 434 with your PR, 126 with trunk. You say that this is better behavior because the work-to-alloc hits the intended target of 4.0. But is that good? I suppose that there is no issue with throughput because each slice doesless work?

(If I understand correctly, the impact on throughput is approximated by the work-to-alloc ratio, so 4 instead of 3.3 means slightly more time spent in the major GC but not much more, and we get to halve the memory usage in return.)