@damiendoligez 2023-04-12
@stedolan @kayceesrk : I think this is ready for review and should be integrated in 5.1, although it's likely that some further improvements are possible. They will be left for a later PR.