@kayceesrk 2023-04-12
The changes are surprisingly simple and look good to me.

One component that I've been wondering about is

> We still need to trigger major slices on other domains when the current domain runs out of things to do. 

We have to do something similar for triggering major slices when the first domain in a minor cycle uses half of its minor heap: https://github.com/ocaml/ocaml/pull/11750. But the mechanisms are different. https://github.com/ocaml/ocaml/pull/11750 interrupts all of the domains by obtaining the `all_domains_lock` and sets the interrupt word to `-1`. See `advance_global_major_slice_epoch` this PR uses an asynchronous version of `caml_try_run_on_all_domains_async` in order to trigger major slice globally. 

I wondered whether both of these can be expressed using a similar mechanism. Have you thought about this @damiendoligez?